% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{apply_denoise}
\alias{apply_denoise}
\title{Apply denoising to a data list}
\usage{
apply_denoise(
  data_list,
  method = "kalman",
  fs = 238,
  bw_cutoff = 15,
  bw_order = 3,
  kalman_dV = NULL,
  kalman_dW = NULL,
  wavelet_name = "d4",
  wavelet_level = 4,
  wavelet_threshold = "soft",
  sg_p = 3,
  sg_n = NULL
)
}
\arguments{
\item{data_list}{List of data items}

\item{method}{Denoising method name}

\item{fs}{Sampling frequency (Hz)}

\item{bw_cutoff}{Butterworth cutoff frequency (Hz)}

\item{bw_order}{Butterworth filter order}

\item{kalman_dV}{Kalman observation noise variance}

\item{kalman_dW}{Kalman state noise variance}

\item{wavelet_name}{Wavelet basis name}

\item{wavelet_level}{Wavelet decomposition level}

\item{wavelet_threshold}{"soft" or "hard"}

\item{sg_p}{Savitzky-Golay polynomial order}

\item{sg_n}{Savitzky-Golay window size}
}
\value{
Modified data_list with denoised Acc_raw and Gyr_raw
}
\description{
Applies the chosen denoising method to all samples in the list.
Supported methods: "none", "butterworth", "kalman", "wavelet", "savgol".
}
