% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFDNxbq-package.R
\docType{data}
\name{BASE_CONFIG}
\alias{BASE_CONFIG}
\title{Default configuration for DFDNxbq experiments}
\format{
A list with the following components:
\describe{
  \item{fs}{Sampling frequency (Hz), default 238}
  \item{sp_ratio}{Stratified sampling ratio, default 0.01}
  \item{fast_subjects}{Subject IDs for LOSO test set}
  \item{L_pre_ms}{Pre-peak window length (ms), default 800}
  \item{delta_ms}{Sliding window delta (ms), default 50}
  \item{enable_dynamic_patching}{Whether to use adaptive patching}
  \item{encoder_type}{One of "cnn", "conv_lstm", "gru_tcn_se"}
  \item{use_hiba}{Whether to use HIBA attention}
  \item{d_model}{Embedding dimension}
  \item{n_epochs}{Training epochs}
  \item{batch_size}{Training batch size}
  \item{denoise_method}{Denoising method: "none", "butterworth", "kalman", "wavelet", "savgol"}
  ... (see source for full list)
}
}
\usage{
BASE_CONFIG
}
\description{
A list of hyperparameters controlling the entire pipeline:
signal processing, deep encoder architecture, HIBA attention,
training, denoising, and classifier settings.
}
\keyword{datasets}
