% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{build_labels}
\alias{build_labels}
\title{Build labels vector from data list}
\usage{
build_labels(data_list)
}
\arguments{
\item{data_list}{Output of \code{load_data} or \code{load_data_from_npz}}
}
\value{
Logical vector, TRUE for fall samples
}
\description{
Build labels vector from data list
}
