% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{build_subjects}
\alias{build_subjects}
\title{Build subject vector from data list}
\usage{
build_subjects(data_list)
}
\arguments{
\item{data_list}{Output of \code{load_data} or \code{load_data_from_npz}}
}
\value{
Integer vector of SubjectIDs
}
\description{
Build subject vector from data list
}
