% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{find_best_threshold}
\alias{find_best_threshold}
\title{Find best threshold on validation set}
\usage{
find_best_threshold(val_true, val_prob, step = 0.01)
}
\arguments{
\item{val_true}{Validation labels}

\item{val_prob}{Validation predicted probabilities}

\item{step}{Threshold search step}
}
\value{
List with threshold and f1
}
\description{
Find best threshold on validation set
}
