% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{load_data}
\alias{load_data}
\title{Load data from an RDS file}
\usage{
load_data(rds_path)
}
\arguments{
\item{rds_path}{Path to the .rds file}
}
\value{
List of data items (each with Acc_raw, Gyr_raw, IsFall, etc.)
}
\description{
Reads a list-of-lists RDS file produced by \code{\link{run_pipeline}} or
\code{\link{load_data_from_npz}}.
}
