% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{predict_catboost}
\alias{predict_catboost}
\title{Predict with a CatBoost model}
\usage{
predict_catboost(model, test_features)
}
\arguments{
\item{model}{catboost model object}

\item{test_features}{Test feature matrix}
}
\value{
List with prob and class elements
}
\description{
Predict with a CatBoost model
}
