% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{predict_rf}
\alias{predict_rf}
\title{Predict with a Random Forest model}
\usage{
predict_rf(model, test_features)
}
\arguments{
\item{model}{randomForest model object}

\item{test_features}{Test feature matrix}
}
\value{
List with class and prob elements
}
\description{
Predict with a Random Forest model
}
