% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{predict_svm}
\alias{predict_svm}
\title{Predict with an SVM model}
\usage{
predict_svm(model, test_features)
}
\arguments{
\item{model}{e1071 svm model object}

\item{test_features}{Test feature matrix}
}
\value{
List with class and prob elements
}
\description{
Predict with an SVM model
}
