% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{stratified_sample}
\alias{stratified_sample}
\title{Stratified sampling from a data list}
\usage{
stratified_sample(data_list, frac = 0.005, seed = 114514)
}
\arguments{
\item{data_list}{List of data items}

\item{frac}{Fraction to sample (default 0.005)}

\item{seed}{Random seed}
}
\value{
Sampled data_list
}
\description{
Preserves class and subject distribution.
}
