% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{train_rf}
\alias{train_rf}
\title{Train a Random Forest classifier}
\usage{
train_rf(
  train_features,
  train_labels,
  ntree = 100,
  mtry = NULL,
  classwt_adl = 1,
  classwt_fall = 3
)
}
\arguments{
\item{train_features}{Training feature matrix}

\item{train_labels}{Training labels (logical)}

\item{ntree}{Number of trees}

\item{mtry}{Number of variables sampled at each split}

\item{classwt_adl}{Weight for ADL class}

\item{classwt_fall}{Weight for Fall class}
}
\value{
A randomForest model object
}
\description{
Train a Random Forest classifier
}
