% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{train_svm}
\alias{train_svm}
\title{Train an SVM classifier}
\usage{
train_svm(
  train_features,
  train_labels,
  cost = 1,
  gamma = NULL,
  classwt_fall = NULL,
  kernel = "linear"
)
}
\arguments{
\item{train_features}{Training feature matrix}

\item{train_labels}{Training labels (logical)}

\item{cost}{SVM cost parameter}

\item{gamma}{SVM gamma parameter}

\item{classwt_fall}{Weight for Fall class (NULL = auto)}

\item{kernel}{SVM kernel type}
}
\value{
An e1071 svm model object
}
\description{
Train an SVM classifier
}
